fix(lock): bypass-proof app lock + listener cleanup (C1, M6)
The PIN lock was bypassable: app/index.tsx called initialize() (loading
keys and starting SPV) BEFORE checking hasPin(); the (tabs) group had no
lock guard so any direct navigation reached authenticated screens; and
the deep-link handler pushed straight to /(tabs)/send regardless of lock
state. The Linking/AppState listeners were also registered in render via
refs with no teardown (M6).

C1:
- Add lock-store.ts: single source of truth for `locked`. Starts locked
  and unresolved so no authenticated screen renders until boot decides.
- Add LockGate (root overlay) rendering LockScreenContent over the whole
  app whenever locked — covers every route, so neither a direct tab
  navigation nor a deep link can reach an authenticated screen. It also
  DEFERS wallet initialization: on a PIN-protected wallet, keys/SPV are
  brought up only after a correct unlock, never behind the lock.
- index.tsx boot no longer initializes before the PIN check; it resolves
  lock state from hasPin() and only initializes when there is no PIN.
- Deep-link handler refuses to navigate while locked or uninitialized.
- Auto-lock now flips the lock store (overlay) instead of navigating to
  a /lock route the user could swipe away from.
- pin-setup unlocks after the user sets their PIN; wipe clears lock
  state; the legacy /lock route redirects to the wallet (overlay is the
  real gate).

M6:
- Linking ("url") and AppState ("change") listeners moved into
  useEffect with subscription.remove() teardown.

Tests: lock-store gate logic — starts locked, opens only on explicit
unlock, re-locks, and the boot decision tracks PIN existence.
