+++ app/index.tsx
@@ -11,10 +11,11 @@ import { useCallback, useState } from "react";
 import { View, Text, ActivityIndicator } from "react-native";
 import { Redirect, useFocusEffect } from "expo-router";
 import { useWalletStore } from "../src/wallet/wallet-store";
+import { useLockStore } from "../src/wallet/lock-store";
 import { getMnemonic, hasPin } from "../src/storage/secure-store";
 import { t } from "../src/i18n";
 
-type AppState = "checking" | "onboarding" | "locked" | "ready" | "error";
+type AppState = "checking" | "onboarding" | "ready" | "error";
 
 export default function IndexScreen() {
   const [appState, setAppState] = useState<AppState>("checking");
@@ -22,6 +23,8 @@ export default function IndexScreen() {
   const hasWallet = useWalletStore((s) => s.hasWallet);
   const initialize = useWalletStore((s) => s.initialize);
   const initialized = useWalletStore((s) => s.initialized);
+  const resolveInitialLock = useLockStore((s) => s.resolveInitialLock);
+  const markNoPinUnlocked = useLockStore((s) => s.markNoPinUnlocked);
 
   useFocusEffect(
     useCallback(() => {
@@ -30,25 +33,29 @@ export default function IndexScreen() {
         try {
           const exists = await hasWallet();
           if (!exists) {
+            // No wallet yet: nothing to lock, go straight to onboarding.
+            markNoPinUnlocked();
             if (!cancelled) setAppState("onboarding");
             return;
           }
 
-          if (!initialized) {
-            const mnemonic = await getMnemonic();
-            if (!mnemonic) {
-              if (!cancelled) setAppState("onboarding");
-              return;
-            }
-            await initialize(mnemonic);
+          const mnemonic = await getMnemonic();
+          if (!mnemonic) {
+            markNoPinUnlocked();
+            if (!cancelled) setAppState("onboarding");
+            return;
           }
 
-          const hasPinSet = await hasPin();
-          if (hasPinSet) {
-            if (!cancelled) setAppState("locked");
-          } else {
-            if (!cancelled) setAppState("ready");
+          // Decide lock state from whether a PIN is set BEFORE exposing keys
+          // (review finding C1). If locked, we render the tabs route but the
+          // root LockGate overlay covers it and only initializes the wallet
+          // after a successful unlock — keys/SPV never come up behind the lock.
+          const pinSet = await hasPin();
+          const willLock = resolveInitialLock(pinSet);
+          if (!willLock && !initialized) {
+            await initialize(mnemonic);
           }
+          if (!cancelled) setAppState("ready");
         } catch (err: unknown) {
           const msg =
             err instanceof Error ? err.message : t("index.error.load");
@@ -62,15 +69,17 @@ export default function IndexScreen() {
       return () => {
         cancelled = true;
       };
-    }, [hasWallet, initialize, initialized]),
+    }, [
+      hasWallet,
+      initialize,
+      initialized,
+      resolveInitialLock,
+      markNoPinUnlocked,
+    ]),
   );
 
   if (appState === "onboarding") {
     return <Redirect href="/onboarding/welcome" />;
-  }
-
-  if (appState === "locked") {
-    return <Redirect href="/lock" />;
   }
 
   if (appState === "ready") {
